+++ src/Attributes/Assertions.php
@@ -1,0 +1,50 @@
+<?php
+
+/*
+ * States.
+ *
+ * LICENSE
+ *
+ * This source file is subject to the 3-Clause BSD license
+ * it is available in LICENSE file at the root of this package
+ * If you did not receive a copy of the license and are unable to
+ * obtain it through the world-wide-web, please send an email
+ * to richard@teknoo.software so we can send you a copy immediately.
+ *
+ *
+ * @copyright   Copyright (c) EIRL Richard Déloge (https://deloge.io - richard@deloge.io)
+ * @copyright   Copyright (c) SASU Teknoo Software (https://teknoo.software - contact@teknoo.software)
+ *
+ * @link        https://teknoo.software/libraries/states Project website
+ *
+ * @license     http://teknoo.software/license/bsd-3         3-Clause BSD License
+ * @author      Richard Déloge <richard@teknoo.software>
+ */
+
+declare(strict_types=1);
+
+namespace Teknoo\States\Attributes;
+
+use Attribute;
+
+/**
+ * Attribute to configure assertions about automation of a stated class.
+ * By default, all assertions attributes are inherited from parent class.
+ *
+ * Usage examples:
+ *   #[StateClass(FooState::class)]
+ *   #[StateClass([FooState::class, BarState::class])]
+ *
+ * @copyright   Copyright (c) EIRL Richard Déloge (https://deloge.io - richard@deloge.io)
+ * @copyright   Copyright (c) SASU Teknoo Software (https://teknoo.software - contact@teknoo.software)
+ * @license     http://teknoo.software/license/bsd-3         3-Clause BSD License
+ * @author      Richard Déloge <richard@teknoo.software>
+ */
+#[Attribute(Attribute::TARGET_CLASS)]
+final class Assertions
+{
+    public function __construct(
+        public readonly bool $inheritsFromParent = true,
+    ) {
+    }
+}